requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Werkzeug
firebase-admin
python-dotenv